+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed `pilet upgrade` command with `npm` client not changing *package.json*
 - Fixed shared dependency list from website emulator to exclude legacy dependencies
 - Added `--allow-self-signed` flag to `piral-cli` commands using HTTP requests
+- Added support for `react-router` v7
 
 ## 1.7.2 (November 8, 2024)